add test, move BootstrapManager to StorageService
