+++ palantir-cassandra-jmx-api/build.gradle
@@ -38,11 +38,3 @@ task copySourceFiles(type: Copy) {
 
 tasks.withType(JavaCompile).all { dependsOn copySourceFiles }
 
-repositories {
-    jcenter()
-}
-
-dependencies {
-    compile 'com.google.guava:guava:18.0'
-}
-
